requirements.txt
amqp==5.3.1
annotated-types==0.7.0
asgiref==3.10.0
attrs==25.4.0
billiard==4.2.2
celery==5.5.3
channels==4.1.0
channels-redis==4.1.0
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
Django==5.1.1
django-cors-headers==4.9.0
django-environ==0.12.0
django-filter==25.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
dotenv==0.9.9
drf-spectacular==0.28.0
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kombu==5.5.4
packaging==25.0
prompt_toolkit==3.0.52
psycopg2==2.9.11
pydantic==2.12.3
pydantic_core==2.41.4
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
PyYAML==6.0.3
redis==5.2.1
referencing==0.37.0
rpds-py==0.28.0
six==1.17.0
sqlparse==0.5.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
vine==5.1.0
wcwidth==0.2.14
requests==2.32.3
weasyprint==60.2
pydyf==0.10.0
flower==2.0.1
daphne==4.1.2
